Textinator/requirements.txt
Babel==2.9.1
beautifulsoup4==4.9.3
Celery==5.2.7
Django==3.2.18
django-clone==3.0.6
django-colorfield==0.4.5
django-compressor==3.0
django-guardian==2.4.0
django-jazzmin==2.4.8
django-maintenancemode==0.11.7
django-modeltranslation==0.17.3
django-nested-admin==3.3.3
django-registration==3.1.2
django-rosetta==0.9.8
django-scientific-survey==0.3.1
django-tinymce4-lite==1.8.0
django_admin_rangefilter==0.6.1
django_filebrowser_no_grappelli==3.8.0
html2markdown==0.1.7
jinja2==3.0.3
jsonlines==3.0.0
Markdown==3.1.1
Pillow==9.4.0
pinax-announcements==4.0.0
psycopg2-binary
pymemcache==3.5.2
PyMySQL==0.9.3
python-magic==0.4.25
pytz==2021.3
pyyaml==6.0
redis==4.2.2
requests==2.22.0
